add user verification

+++ views/doctor.py
@@ -1,11 +1,12 @@
 from datetime import datetime
-from telnetlib import STATUS
 from schemas import educations_schema, education_schema, doctors_schema, doctor_schema
 from flask import Blueprint, jsonify, request, Response
 from models.doctor import Doctor, Education
 from config import db
-from services.auth import token_required, get_user_from_token
-from services.doctor import get_doctor_by_id, get_education_by_id_from_doctor
+from services.auth import token_required, get_user_from_request
+from services.doctor import get_doctor_by_id,\
+     get_education_by_id_from_doctor,\
+     is_equal_user_from_request_with_user_db        
 from constants import DATE_FORMAT
 
 
@@ -41,7 +42,7 @@ def get_doctors():
 @token_required
 @doctor_view.route("/doctor", methods=["Post"])
 def create_doctor():
-    user = get_user_from_token(request)
+    user = get_user_from_request(request)
     new_doctor = Doctor(
         first_name = request.json["first_name"],
         surname = request.json["surname"],
@@ -53,8 +54,9 @@ def create_doctor():
     else:
         user.doctor = new_doctor
     db.session.add(new_doctor)
-    db.session.add(user)
     db.session.commit()
+
+    return Response(jsonify(doctor_schema.dump(new_doctor)), status=201, mimetype='application/json')
 
 
 @doctor_view.route("/doctor/<id>", methods=["Get"])
@@ -62,29 +64,37 @@ def get_doctor(id):
     doctor = get_doctor_by_id(id)
     return doctor_schema.dump(doctor)
 
-
+@token_required
 @doctor_view.route("/doctor/<id>", methods=["Put"])
 def put_doctor(id):
+    user = get_user_from_request(request)
     doctor = get_doctor_by_id(id)
+    if not is_equal_user_from_request_with_user_db(request_user=user, db_user=doctor.user):
+        return Response(status=403)
     doctor.first_name = request.json["first_name"]
     doctor.surname = request.json["surname"]
     doctor.birthday = datetime.strptime(request.json["birthday"], DATE_FORMAT)
-    education = get_education_from_request(request.json["education"])
     db.session.commit()
     return Response(status=200)
 
-
+@token_required
 @doctor_view.route("/doctor/<id>", methods=["Delete"])
 def delete_doctor(id):
+    user = get_user_from_request(request)
     doctor = get_doctor_by_id(id)
+    if not is_equal_user_from_request_with_user_db(request_user=user, db_user=doctor.user):
+        return Response(status=403)
     db.session.delete(doctor)
     db.session.commit()
     return Response(status=204)
 
-
+@token_required
 @doctor_view.route("/doctor/<doctor_id>/education", methods=["Post"])
 def create_education(doctor_id):
+    user = get_user_from_request(request)
     doctor = get_doctor_by_id(doctor_id)
+    if not is_equal_user_from_request_with_user_db(request_user=user, db_user=doctor.user):
+        return Response(status=403)
     doctor.education = get_education_from_request(request.json)
     db.session.commit()
 
@@ -101,9 +111,13 @@ def get_education_by_id(doctor_id, education_id):
     education = get_education_by_id_from_doctor(doctor, education_id)
     return education_schema.dump(education)
 
+@token_required
 @doctor_view.route("/doctor/<doctor_id>/education/<education_id>", methods=["Put"])
 def update_education_by_id(doctor_id, education_id):
-    doctor = get_doctor_by_id(doctor_id)
+    user = get_user_from_request(request)
+    doctor = get_doctor_by_id(id)
+    if not is_equal_user_from_request_with_user_db(request_user=user, db_user=doctor.user):
+        return Response(status=403)
     education = get_education_by_id_from_doctor(doctor, education_id)
     education.university_name = request.json['university_name']
     education.specialization = request.json['specialization']
@@ -112,9 +126,13 @@ def update_education_by_id(doctor_id, education_id):
     db.session.commit()
     return Response(status=200)
 
+@token_required
 @doctor_view.route("/doctor/<doctor_id>/education/<education_id>", methods=["Delete"])
 def delete_education_by_id(doctor_id, education_id):
-    doctor = get_doctor_by_id(doctor_id)
+    user = get_user_from_request(request)
+    doctor = get_doctor_by_id(id)
+    if not is_equal_user_from_request_with_user_db(request_user=user, db_user=doctor.user):
+        return Response(status=403)
     education = get_education_by_id_from_doctor(doctor, education_id)
     db.session.delete(education)
     db.session.commit()
